requirements.txt
alabaster==0.7.12
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
black==20.8b1
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
Click
cloudpickle==1.6.0
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
docutils==0.16
entrypoints==0.3
flake8==3.8.4
Flask==1.1.1
Flask-Cors==3.0.10
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
helpdev==0.7.1
idna==2.10
imagesize==1.2.0
intervaltree==3.1.0
ipykernel==5.3.4
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.5.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
keyring==21.5.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.2.2
marshmallow-sqlalchemy==0.24.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
numpy==1.20.0
numpydoc==1.1.0
packaging==20.4
pandas==1.2.1
pandocfilters==1.4.3
parso==0.7.0
pathspec==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
psutil==5.7.3
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.6.0
pyls-black==0.4.6
pyls-spyder==0.1.1
PyMySQL==1.0.2
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.1
pytz==2019.3
pyxdg==0.27
pyzmq==20.0.0
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.11.13
requests==2.25.0
rope==0.18.0
SecretStorage==3.2.0
six==1.13.0
snowballstemmer==2.0.0
sortedcontainers==2.3.0
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.2.0
spyder-kernels==1.10.0
SQLAlchemy==1.3.10
testpath==0.4.4
three-merge==0.1.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.12.1
wurlitzer==2.0.1
yapf==0.30.0
